restructed triangle count
