+++ docbrown/db/src/algorithms/mod.rs
@@ -1,7 +1,6 @@
 pub mod degree;
 pub mod directed_graph_density;
-pub mod global_triangle_count;
 pub mod local_clustering_coefficient;
-pub mod local_triangle_count;
+pub mod triangle_count;
 pub mod reciprocity;
 pub mod connected_components;
