+++ examples/rust/src/bin/hulongbay/main.rs
@@ -13,7 +13,7 @@ use docbrown_core::tgraph::TemporalGraph;
 use docbrown_core::{state, utils};
 use docbrown_core::{Direction, Prop};
 use docbrown_db::algorithms::connected_components::weakly_connected_components;
-use docbrown_db::algorithms::global_triangle_count::global_triangle_count;
+use docbrown_db::algorithms::triangle_count::global_triangle_count;
 use docbrown_db::csv_loader::csv::CsvLoader;
 use docbrown_db::program::algo::triangle_counting_fast;
 use docbrown_db::program::{
